crypto_etl/requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
pytest